Refactors to use more enums
