+++ lib/clientHelpers.ts
@@ -9,6 +9,7 @@ import {
   PublicFormRecord,
   Response,
   Responses,
+  VaultStatus,
 } from "@lib/types";
 import { Submission } from "@lib/types/submission-types";
 import { getCsrfToken } from "next-auth/react";
@@ -349,4 +350,14 @@ export const ucfirst = (string: string) => {
     return "";
   }
   return string.charAt(0).toUpperCase() + string.slice(1);
+};
+
+export const isStatus = (query: string, status: VaultStatus | VaultStatus[]): boolean => {
+  const ucQuery = ucfirst(query);
+  if (Array.isArray(status)) {
+    return [VaultStatus.NEW, VaultStatus.DOWNLOADED, VaultStatus.CONFIRMED].includes(
+      ucQuery as VaultStatus
+    );
+  }
+  return ucQuery === status;
 };
